11-JL-SecurityConfigurationFix

+++ backend/src/main/java/es/dawgroup2/juding/users/rest/AdminUserAPIController.java
@@ -1,10 +1,10 @@
 package es.dawgroup2.juding.users.rest;
 
+import com.fasterxml.jackson.annotation.JsonView;
 import es.dawgroup2.juding.auxTypes.belts.BeltService;
 import es.dawgroup2.juding.auxTypes.gender.GenderService;
 import es.dawgroup2.juding.auxTypes.refereeRange.RefereeRangeService;
 import es.dawgroup2.juding.main.DateService;
-import es.dawgroup2.juding.main.HeaderInflater;
 import es.dawgroup2.juding.main.image.ImageService;
 import es.dawgroup2.juding.users.User;
 import es.dawgroup2.juding.users.UserService;
@@ -23,14 +23,9 @@ import org.springframework.web.bind.annotation.*;
 import javax.validation.Valid;
 import java.util.List;
 
-import static org.springframework.web.servlet.support.ServletUriComponentsBuilder.fromCurrentRequest;
-
 @RestController
-@RequestMapping("/api/admin/user")
+@RequestMapping("/api")
 public class AdminUserAPIController {
-    @Autowired
-    HeaderInflater headerInflater;
-
     @Autowired
     UserService userService;
 
@@ -116,55 +111,13 @@ public class AdminUserAPIController {
         int defPage = (page == null) ? 1 : page;
         if (defPage < 0) return ResponseEntity.badRequest().build();
         Page<User> requiredPage = userService.getActiveRefereesInPages(defPage);
-        if (requiredPage.hasContent())
-            return ResponseEntity.ok(requiredPage);
-        else
-            return ResponseEntity.badRequest().build();
-    }
-
-    /**
-     * Saves new information associated with a user and returns the user with the new information modified.
-     *
-     * @param adminUserEditionDTO Admin user edition Data Transfer Object.
-     * @return User object with new information saved.
-     */
-    @Operation(summary = "Saves new information associated with a user and returns the user with the new information modified.")
-    @ApiResponses(value = {
-            @ApiResponse(responseCode = "201", description = "User object with new information saved.",
-                    content = {@Content(mediaType = "application/json",
-                            schema = @Schema(implementation = User.class))}),
-            @ApiResponse(responseCode = "403", description = "Not allowed (there is not logged in user or it is not an administrator).",
-                    content = @Content)
-    })
-    @PutMapping("/")
-    public ResponseEntity<User> savingUser(@Valid @Parameter(description = "Admin user edition Data Transfer Object.") @RequestBody AdminUserEditionDTO adminUserEditionDTO) {
-        if (userService.matchingLicenseAndNickname(adminUserEditionDTO.getLicenseId(), adminUserEditionDTO.getNickname())) {
-            return ResponseEntity.badRequest().build();
-        }
-        User user = userService.save(adminUserEditionDTO.getName(),
-                adminUserEditionDTO.getSurname(),
-                adminUserEditionDTO.getGender(),
-                adminUserEditionDTO.getPhone(),
-                adminUserEditionDTO.getEmail(),
-                adminUserEditionDTO.getBirthDate(),
-                adminUserEditionDTO.getDni(),
-                adminUserEditionDTO.getLicenseId(),
-                adminUserEditionDTO.getNickname(),
-                null,
-                null,
-                null,
-                null,
-                adminUserEditionDTO.getBelt(),
-                null, adminUserEditionDTO.getGym(),
-                adminUserEditionDTO.getWeight(),
-                adminUserEditionDTO.getRefereeRange());
-        return ResponseEntity.created(fromCurrentRequest().path("/api/admin/user/{licenseId}").buildAndExpand(user.getLicenseId()).toUri()).body(user);
+        return (requiredPage.hasContent()) ? ResponseEntity.ok(requiredPage) : ResponseEntity.badRequest().build();
     }
 
     /**
      * Process a referee's application and saves it as an official referee.
      *
-     * @param licenseId License ID of admitted referee.
+     * @param id License ID of admitted referee.
      * @return New referee's complete information.
      */
     @Operation(summary = "Process a referee's application and saves it as an official referee.")
@@ -172,21 +125,21 @@ public class AdminUserAPIController {
             @ApiResponse(responseCode = "200", description = "New referee's complete information.",
                     content = {@Content(mediaType = "application/json",
                             schema = @Schema(implementation = User.class))}),
-            @ApiResponse(responseCode = "400", description = "Requested license ID is not registered in the application or referee has been admitted before",
+            @ApiResponse(responseCode = "400", description = "Requested license ID is not registered in the application or referee has been admitted before.",
                     content = @Content),
             @ApiResponse(responseCode = "403", description = "Not allowed (there is not logged in user or it is not an administrator).",
                     content = @Content)
     })
-    @PutMapping("/refereeApplication/{licenseId}")
-    public ResponseEntity<User> admitReferee(@Parameter(description = "License ID of admitted referee.") @PathVariable String licenseId) {
-        User user = userService.admitReferee(licenseId);
+    @PutMapping("/referees/applications/{id}")
+    public ResponseEntity<User> admitReferee(@Parameter(description = "License ID of admitted referee.") @PathVariable String id) {
+        User user = userService.admitReferee(id);
         return (user == null) ? ResponseEntity.badRequest().build() : ResponseEntity.ok(user);
     }
 
     /**
      * Deletes a user (only successful if user has not taken part of any competition previously or if it has joined no one).
      *
-     * @param licenseId License ID of user.
+     * @param id License ID of user.
      * @return Deleted user information.
      */
     @Operation(summary = "Deletes a user.", description = "Warning: this method is only successful if user has not taken part of any competition previously or if it has joined no one.")
@@ -196,13 +149,14 @@ public class AdminUserAPIController {
                             schema = @Schema(implementation = User.class))}),
             @ApiResponse(responseCode = "403", description = "Not allowed (there is not logged in user or it is not an administrator).",
                     content = @Content),
-            @ApiResponse(responseCode = "404", description = "Not able to delete user (user did not exist or user has taken part of any past or future competition)",
+            @ApiResponse(responseCode = "404", description = "Not able to delete user (user did not exist or user has taken part of any past or future competition).",
                     content = @Content)
     })
-    @DeleteMapping("/{licenseId}")
-    public ResponseEntity<User> deleteUser(@Parameter(description = "License ID of user.") @PathVariable String licenseId) {
-        User user = userService.getUserOrNull(licenseId);
+    @DeleteMapping(value = {"/competitors/{id}", "/referees/{id}"})
+    public ResponseEntity<User> deleteUser(@Parameter(description = "License ID of user.") @PathVariable String id) {
+        User user = userService.getUserOrNull(id);
+        if (user == null) return ResponseEntity.notFound().build();
         userService.delete(user);
-        return (user == null) ? ResponseEntity.notFound().build() : ResponseEntity.ok(user);
+        return ResponseEntity.ok(user);
     }
 }
